+++ lib/src/postgres_api.dart
@@ -7,41 +7,53 @@ class PostgresApi extends DatabaseApi {
   PostgresApi(this._connection);
 
   @override
-  Future<Iterable<String>> getTables() async => (await query('''
-    SELECT table_name FROM information_schema.tables
-    WHERE table_type='BASE TABLE' AND table_schema='public'
-  ''')).map((e) => e['table_name'] as String?).whereType<String>();
-
-  @override
-  Future<Iterable<String>> getPrimaryKeys(String table) async =>
-      (await query('''
-        SELECT a.attname AS name
-        FROM
-          pg_class AS c
-          JOIN pg_index AS i ON c.oid = i.indrelid AND i.indisprimary
-          JOIN pg_attribute AS a ON c.oid = a.attrelid AND a.attnum = ANY(i.indkey)
-        WHERE c.oid = ?1::regclass
-      ''', [table])).map((e) => e['name'] as String);
-
-  @override
   Future<void> execute(String sql, [List<Object?>? args]) =>
-      _connection.execute(
-        Sql.indexed(sql, substitution: '?'),
-        parameters: args,
-        ignoreRows: true,
-      );
+      _connection.execute(sql.asQuery, parameters: args, ignoreRows: true);
 
   @override
   Future<List<Map<String, Object?>>> query(String sql,
           [List<Object?>? args]) async =>
-      (await _connection.execute(
-        Sql.indexed(sql, substitution: '?'),
-        parameters: args,
-      ))
+      (await _connection.execute(sql.asQuery, parameters: args))
           .map((e) => e.toColumnMap())
           .toList();
 
   @override
-  Future<void> transaction(Future<void> Function(DatabaseApi txn) action) =>
-      (_connection as SessionExecutor).runTx((t) => action(PostgresApi(t)));
+  Future<void> transaction(Future<void> Function(DatabaseApi txn) actions) =>
+      (_connection as SessionExecutor).runTx((t) => actions(PostgresApi(t)));
+
+  @override
+  Future<void> executeBatch(Future<void> Function(WriteApi api) actions) async {
+    await (_connection as SessionExecutor).runTx((t) async {
+      final batch = _BatchApi(t);
+      await actions(batch);
+      await batch.commit();
+    });
+  }
+}
+
+class _BatchApi extends WriteApi {
+  final Session _connection;
+  String? _batchedSql;
+  Statement? _batch;
+
+  _BatchApi(this._connection);
+
+  @override
+  Future<void> execute(String sql, [List<Object?>? args]) async {
+    if (sql != _batchedSql) {
+      await _batch?.dispose();
+      _batch = await _connection.prepare(sql.asQuery);
+      _batchedSql = sql;
+    }
+
+    await _batch!.run(args);
+  }
+
+  Future<void> commit() async {
+    await _batch!.dispose();
+  }
+}
+
+extension on String {
+  Sql get asQuery => Sql.indexed(this, substitution: '?');
 }
